2011-01-28  Alejandro G. Castro  <alex@igalia.com>
        Reviewed by Xan Lopez.

        [GTK] Fix compilation warning
        https://bugs.webkit.org/show_bug.cgi?id=53292

        * webkit/webkitwebframe.cpp:
        (draw_page_callback): Cast the pageCount from size_t to int, it
        should be safe considering page_nr is the number of pages,
        0-based.

Canonical link: https://commits.webkit.org/67112@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@76940 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/gtk/webkit/webkitwebframe.cpp
@@ -768,7 +768,7 @@ static void draw_page_callback(GtkPrintOperation* op, GtkPrintContext* context, 
 {
     PrintContext* printContext = reinterpret_cast<PrintContext*>(user_data);
 
-    if (page_nr >= printContext->pageCount())
+    if (page_nr >= static_cast<gint>(printContext->pageCount()))
         return;
 
     cairo_t* cr = gtk_print_context_get_cairo_context(context);
